reduce cluster number
