--- integration-test/src/test/java/org/apache/iotdb/ainode/it/AIEnvIT.java
@@ -1,86 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-package org.apache.iotdb.ainode.it;
-
-import org.apache.iotdb.it.env.EnvFactory;
-import org.apache.iotdb.it.framework.IoTDBTestRunner;
-import org.apache.iotdb.itbase.category.AIClusterIT;
-
-import org.junit.After;
-import org.junit.Before;
-import org.junit.Test;
-import org.junit.experimental.categories.Category;
-import org.junit.runner.RunWith;
-
-import java.sql.Connection;
-import java.sql.ResultSet;
-import java.sql.ResultSetMetaData;
-import java.sql.SQLException;
-import java.sql.Statement;
-
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.fail;
-
-@RunWith(IoTDBTestRunner.class)
-@Category({AIClusterIT.class})
-public class AIEnvIT {
-  @Before
-  public void setUp() throws Exception {
-    // Init 1C1D1M cluster environment
-    EnvFactory.getEnv().initClusterEnvironment(1, 1);
-  }
-
-  @After
-  public void tearDown() {
-    EnvFactory.getEnv().cleanClusterEnvironment();
-  }
-
-  private static void checkHeader(ResultSetMetaData resultSetMetaData, String title)
-      throws SQLException {
-    String[] headers = title.split(",");
-    for (int i = 1; i <= resultSetMetaData.getColumnCount(); i++) {
-      assertEquals(headers[i - 1], resultSetMetaData.getColumnName(i));
-    }
-  }
-
-  @Test
-  public void aiNodeConnectionTest() {
-    String sql = "SHOW AINODES";
-    String title = "NodeID,Status,RpcAddress,RpcPort";
-    try (Connection connection = EnvFactory.getEnv().getConnection();
-        Statement statement = connection.createStatement()) {
-
-      try (ResultSet resultSet = statement.executeQuery(sql)) {
-        ResultSetMetaData resultSetMetaData = resultSet.getMetaData();
-        checkHeader(resultSetMetaData, title);
-        int count = 0;
-        while (resultSet.next()) {
-          assertEquals("2", resultSet.getString(1));
-          assertEquals("Running", resultSet.getString(2));
-          count++;
-        }
-        assertEquals(1, count);
-      }
-    } catch (SQLException e) {
-      e.printStackTrace();
-      fail(e.getMessage());
-    }
-  }
-}
